+++ src/OpenTelemetry.AutoInstrumentation/Constants.cs
@@ -21,6 +21,7 @@ internal static class Constants
     public static class Tracer
     {
         public const string Version = "0.3.1";
+        public const string AutoInstrumentationVersionName = "telemetry.auto.version";
     }
 
     public static class ConfigurationValues